requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.21.0
attrs==21.2.0
Babel==2.9.1
certifi==2020.12.5
chardet==4.0.0
environs==9.3.2
gino==1.0.1
hiredis==2.0.0
idna==3.1
marshmallow==3.12.1
multidict==5.1.0
python-dotenv==0.17.1
pytz==2021.1
SQLAlchemy==1.3.24
typing-extensions==3.10.0.0
yarl==1.6.3

Django~=3.2.4
django-jsonfield~=1.4.1
asgiref~=3.3.4